Import vendor/NetBSD/tests@b34f66deb
This change imports the latest snapshot of vendor/NetBSD/tests:
b34f66deb.

All code has been rebased or marked as an expected error to ensure that
the tests run cleanly on all Tier 1 architectures.

Any modifications that were made and were not upstreamed which were
specific to FreeBSD and were not easily mergeable have been removed and
placed in their own FreeBSD-specific tests to ease future merges.

Any and all FreeBSD-specific tests should be written as separate tests
expressed using either the ATF or GoogleTest frameworks.

+++ contrib/netbsd-tests/crypto/libcrypto/t_libcrypto.sh
@@ -1,4 +1,4 @@
-# $NetBSD: t_libcrypto.sh,v 1.4 2016/10/13 09:25:37 martin Exp $
+# $NetBSD: t_libcrypto.sh,v 1.9 2023/05/08 17:46:06 christos Exp $
 #
 # Copyright (c) 2008, 2009, 2010 The NetBSD Foundation, Inc.
 # All rights reserved.
@@ -32,24 +32,15 @@ engine_head()
 }
 engine_body()
 {
-	atf_check -o ignore -e ignore "$(atf_get_srcdir)/h_enginetest"
-}
-
-atf_test_case rand
-rand_head()
-{
-	atf_set "descr" "Checks peudo-random number generator"
-}
-rand_body()
-{
-	atf_check -o ignore -e ignore "$(atf_get_srcdir)/h_randtest"
+	atf_check -o ignore -e ignore "$(atf_get_srcdir)/h_enginetest" \
+	    "$(atf_get_srcdir)/d_server.pem"
 }
 
 atf_test_case bn
 bn_head()
 {
 	atf_set "descr" "Checks BIGNUM library"
-	atf_set "timeout" "360"
+	atf_set "timeout" "7200"
 }
 bn_body()
 {
@@ -71,17 +62,6 @@ conf_body()
 		$(atf_get_srcdir)/h_conftest
 }
 
-atf_test_case lhash
-lhash_head()
-{
-	atf_set "descr" "Checks lhash - dynamic hash tables"
-}
-lhash_body()
-{
-	atf_check -o ignore -e ignore -x \
-		"echo hoge | $(atf_get_srcdir)/h_lhashtest"
-}
-
 atf_test_case threads
 threads_head()
 {
@@ -89,19 +69,23 @@ threads_head()
 }
 threads_body()
 {
-	$(atf_get_srcdir)/h_threadstest \
-	    -cert $(atf_get_srcdir)/d_server.pem \
-	    -ccert $(atf_get_srcdir)/d_client.pem \
-	    2>&1 | tee out
-	atf_check -s eq:1 -o empty -e empty grep :error: out
+	local s=$(atf_get_srcdir)
+	if [ -f "$s/rsakey.pem" ]; then
+	    atf_check -o ignore -e ignore "$s/h_threadstest" \
+		-config "$s/default.cnf" "$s"
+	else
+		"$s/h_threadstest" \
+		    -cert "$s/d_server.pem" \
+		    -ccert "$s/d_client.pem" \
+		2>&1 | tee out
+		atf_check -s eq:1 -o empty -e empty grep :error: out
+	fi
 }
 
 atf_init_test_cases()
 {
 	atf_add_test_case engine
-	atf_add_test_case rand
 	atf_add_test_case bn
 	atf_add_test_case conf
-	atf_add_test_case lhash
 	atf_add_test_case threads
 }
